dom/nodes/insertion-effects/Node-appendChild-iframe-before-script.tentative.html

<!doctype html>
<meta charset=utf-8>
<title>Node.appendChild: inserting script and iframe from a div</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id="log"></div>
<script>
let state = "script not run yet";
const iframe = document.createElement("iframe");

test(() => {
  iframe.src = "data:text/html,";

  const script = document.createElement("script");
  script.textContent = `
    state = iframe.contentWindow ? "iframe with content window" : "iframe without content window";
`;

  const div = document.createElement("div");
  div.appendChild(iframe);
  div.appendChild(script);

  assert_equals(state, "script not run yet");
  document.body.appendChild(div);
  assert_equals(state, "iframe with content window");
});
</script>
